空「.class」ファイルの生成

+++ forgive/tempfile/TempFileLapper.java
@@ -1,10 +1,14 @@
 package forgive.tempfile;
 
+import java.io.BufferedInputStream;
+import java.io.BufferedOutputStream;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
-import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.StandardOpenOption;
 import java.util.EnumMap;
 import java.util.Map;
 
@@ -12,10 +16,19 @@ public class TempFileLapper {
     private Map<TempFiles, Path> files = new EnumMap<>(TempFiles.class);
 
     public void createAllTempFile(){
+        createAllTempFile(true);
+    }
+
+    /**
+     * 基本的にデバッグ用途
+     */
+    public void createAllTempFile(boolean deleteOnClose){
         try{
             for(TempFiles fileEnum:TempFiles.values()){
                 Path path = Files.createTempFile(fileEnum.fileName, null);
-                path.toFile().deleteOnExit();
+                if (deleteOnClose){
+                    path.toFile().deleteOnExit();
+                }
                 files.put(fileEnum, path);
             }
         }catch(Exception e){
@@ -23,9 +36,14 @@ public class TempFileLapper {
         }
     }
 
+    /**
+     * Appendモードで書き込みます
+     */
     public BufferedWriter getWriter(TempFiles file){
         try{
-            return Files.newBufferedWriter(files.get(file));
+            return Files.newBufferedWriter(files.get(file), StandardOpenOption.CREATE,
+                                                            StandardOpenOption.WRITE,
+                                                            StandardOpenOption.APPEND);
         }catch(Exception e){
             throw new RuntimeException("一時ファイル読み込みに失敗しました。");
         }
@@ -39,8 +57,30 @@ public class TempFileLapper {
         }
     }
 
+    /**
+     * Appendモードで書き込みます
+     */
+    public OutputStream getOutputStream(TempFiles file){
+        try{
+            return new BufferedOutputStream(Files.newOutputStream(files.get(file),
+                                            StandardOpenOption.CREATE, 
+                                            StandardOpenOption.WRITE,
+                                            StandardOpenOption.APPEND));
+        }catch(Exception e){
+            throw new RuntimeException("一時ファイル読み込みに失敗しました。");
+        }
+    }
+
+    public InputStream getInputStream(TempFiles file){
+        try{
+            return new BufferedInputStream(Files.newInputStream(files.get(file)));
+        }catch(Exception e){
+            throw new RuntimeException("一時ファイル書き込みに失敗しました。");
+        }
+    }
+
     public enum TempFiles{
-        SRC_FILE_SEPARATE("src_s")
+        SRC_FILE_SEPARATE("src_s"), OUTPUT_CLASSFILE("out")
         ;
         private final String fileName;
         TempFiles(String fileName){
